+++ src/decisions/admin/decisions.controller.spec.ts
@@ -19,6 +19,7 @@ import { DecisionDatasetsPresenter } from './presenters/decision-datasets.presen
 import * as checkCanChangeDatasetModule from './helpers/check-can-change-dataset.helper';
 import { DecisionsCsvWriter } from './helpers/decisions-csv-writer.helper';
 import { Response } from 'express';
+import { DecisionDatasetStatus } from '../decision-dataset.entity';
 
 jest.mock('./presenters/decision-datasets.presenter');
 jest.mock('../presenters/decision-dataset.presenter');
@@ -226,6 +227,7 @@ describe('DecisionsController', () => {
         profession: profession,
         organisation: organisation,
         year: 2017,
+        status: DecisionDatasetStatus.Live,
       });
 
       const request = createDefaultMockRequest();
@@ -246,6 +248,7 @@ describe('DecisionsController', () => {
         organisation,
         tables: mockTables,
         year: 2017,
+        isPublished: true,
       };
 
       const result = await controller.show(
@@ -279,6 +282,46 @@ describe('DecisionsController', () => {
         i18nService,
       );
       expect(DecisionDatasetPresenter.prototype.tables).toHaveBeenCalled();
+    });
+
+    describe('when the dataset has not yet been published', () => {
+      it('sets `isPublished` to false', async () => {
+        const profession = professionFactory.build({
+          id: 'example-profession-id',
+        });
+        const organisation = organisationFactory.build({
+          id: 'example-organisation-id',
+        });
+
+        const dataset = decisionDatasetFactory.build({
+          profession: profession,
+          organisation: organisation,
+          year: 2017,
+          status: DecisionDatasetStatus.Draft,
+        });
+
+        const request = createDefaultMockRequest();
+
+        jest
+          .spyOn(checkCanChangeDatasetModule, 'checkCanChangeDataset')
+          .mockImplementation();
+
+        professionsService.findWithVersions.mockResolvedValueOnce(profession);
+        decisionDatasetsService.find.mockResolvedValue(dataset);
+
+        (
+          DecisionDatasetPresenter.prototype.tables as jest.Mock
+        ).mockReturnValue(mockTables);
+
+        const result = await controller.show(
+          'example-profession-id',
+          'example-organisation-id',
+          2017,
+          request,
+        );
+
+        expect(result.isPublished).toEqual(false);
+      });
     });
   });