+++ src/decisions/admin/interfaces/publication-template.interface.ts
@@ -1,11 +1,8 @@
-import { Table } from '../../../common/interfaces/table';
 import { Organisation } from '../../../organisations/organisation.entity';
 import { Profession } from '../../../professions/profession.entity';
 
-export interface ShowTemplate {
+export interface PublicationTemplate {
   profession: Profession;
   organisation: Organisation;
   year: number;
-  tables: Table[];
-  isPublished: boolean;
 }